+++ src/domible/builders/__init__.py
@@ -3,5 +3,6 @@
 from domible.builders.tableBuilder import RowBuilder, TableBuilder, build_table_from_dicts
 from domible.builders.navBuilder import NavBuilder 
 from domible.builders.elementFromObject import element_from_object  
+from domible.builders.formBuilder import ToggleDetailsButton, default_toggle_details_button  
 
 # end of file