+++ docs/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## v0.1.15 (2025-01-13) 
+
+- this version is motivated by a desire to toggle visibility of all details elements 
+- and now, forms... starting with the button  
+  - this is a very basic ability to create a button element
+  - any attributes (e.g., type) will be done manually or via a builders class
+  - this is all TBD and might change in subsequent releases 
+- a new builder to add a toggle button to a page to expand/collapse the content of a details element 
+  - this is also the introduction of JavaScript into domible (sorry, it was inevitable) 
+  - the JS is hard coded into the builder, which might be the long term solution, but I doubt it 
+  - focus is on functionality, no CSS has been added 
+- added toggle details button into efo.py in tests 
+- added a toggle details button in open_object_in_browser() in domible tools 
+- fixed bug in Script element where contents was not being passed to super().__init__ 
+
 ## v0.1.14 (2024-12-19) -- BREAKING BACKWARD COMPATIBILITY 
 
 - removed typer from dicli, starts a little faster